Describe the segment_size and counter keyword arguments. Another bunch of rewrites and clarifications.
I now think this PEP is complete enough to implement.  My next step
will be to implement it and see if any parts turn out to be horribly
complicated or inefficient.  If nothing turns up, I'll finalize the
PEP; if there are problems, the PEP will be revised in light of
implementation experience